Change the method in which the retrieval process functions
Instead of getting the itemid from the item table it will now refer to the borrow table

The system will now automatically know if the item is late or not. No need for manual input.

All that's left to do is implement an input for fines for each scanned item.